Some polish, narrative and additional analysis.
This adds a bit of polish and narrative. I reformatted the HTML
a bit, and added more copy. The goal of which was to both enforce
the motivations and take the user on a bit of journey. I'm sure
we'll make some changes to it as we refine things further.

I also added code for determining the intersection of the author's
citations. This produced a pretty interesting list. Though there's
not as much overlap as I'd hoped. This might increase if we start
with more papers, rather than truncating to the top 10.

I left a placeholder for some sort of similarity analysis. I think
we might be able to use cosine similarity and a graph to cluster authors
based on the SPECTRE embeddings of their papers. I've already fetched
the embeddings from the API, so we just have to wire up the mechanism
for calculating cosine similarity and outputting a plot.